Rename Control.Algebra.Internal to Control.Effect.Internal.
